% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauli_table.R
\name{paulitable}
\alias{paulitable}
\title{Get variable names, labels, and optional summary stats}
\usage{
paulitable(data, stats = FALSE, missings = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{stats}{Logical, whether to include mean and range (for numeric vars)}

\item{missings}{Logical, whether to include missing value percentage}
}
\value{
A data frame with variable names, labels, and optional stats
}
\description{
RStudio unlike STATA has not a Variable-Window which quickly
lets you see the variables and corresponding variable labels.
Therefore,  paulitable() let's you create a dataframe of variable names and their variable labels,
with optional summary statistics like mean, range, and missing percentage.
}
